app/core_api/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9  # For PostgreSQL database
gunicorn>=21.2.0  # For production server
whitenoise>=6.6.0  # For serving static files
pillow>=10.0.0  # For image processing
django-filter>=23.5  # For filtering in DRF
requests>=2.31.0  # For HTTP requests
drf-spectacular>=0.27.0  # For API documentation
